% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akcoast.R
\docType{data}
\name{akcoast}
\alias{akcoast}
\title{Alaska coastline}
\format{
Simple feature collection with 273 features and 5 fields:
\describe{
\item{geometry}{POLYGON}
}
}
\source{
\url{https://gis.data.alaska.gov/datasets/alaska-1250000}
}
\usage{
akcoast
}
\description{
Alaska 1:250000 coastal data polygon. This is provided by the Alaska Department of
Natural Resources and was obtained from their open data portal
(https://gis.data.alaska.gov/datasets/alaska-1250000). Note, only those polygons that
intersect with the bounding box of our harbor seal movement are included.
}
\keyword{datasets}
