% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_barrier_segments.R
\name{get_barrier_segments}
\alias{get_barrier_segments}
\title{Identify track points that intersect with a barrier polygon}
\usage{
get_barrier_segments(trkpts, barrier)
}
\arguments{
\item{trkpts}{Simple Feature points ('sf', 'sfc_POINT'/'sfc_MULTIPOINT') that represent
track points. Order is accepted as is and the bounding box of trkpts should be within
the bounding box of the barrier polygon.}

\item{barrier}{Simple Feature polygon ('sf', 'sfc_POLYGON'/'sfc_MULTIPOLYGON')
representing the barrier feature. Should be the same barrier as supplied to the
\code{prt_visgraph()} function.}
}
\value{
tibble representing segments of consecutive points that intersect with
the barrier feature. the \emph{start_pt} and \emph{end_pt} geometry columns represent the bookend
points for each segment that do not intersect with the barrier feature. The \emph{n_pts}
column is the number of points to be re-routed.
}
\description{
This function identifies the segments of consecutive points that intersect with the
barrier polygon feature. The result is a tibble of segment records that identify
portions of the track that will need to be re-routed.
}
