% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_barrier.R
\docType{data}
\name{land_barrier}
\alias{land_barrier}
\title{land barrier}
\format{
Simple feature collection with 19 features and 0 fields:
\describe{
\item{geometry}{MULTIPOLYGON}
}
}
\source{
geopackage file available in extData
}
\usage{
land_barrier
}
\description{
A polygon dataset used to test and demonstrate package functions for routing
paths around barriers
}
\keyword{datasets}
