% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poi.R
\docType{data}
\name{poi}
\alias{poi}
\title{points of interest}
\format{
Simple feature collection with 67 features and 0 fields:
\describe{
\item{geometry}{MULTIPOINT}
}
}
\source{
geopackage file available in extData
}
\usage{
poi
}
\description{
A point dataset used to test and demonstrate package functions for routing
paths around barriers
}
\keyword{datasets}
