% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_extend_path.R
\name{prt_extend_path}
\alias{prt_extend_path}
\title{Extend a path to include given start and end points}
\usage{
prt_extend_path(l_geom, start_pt, end_pt)
}
\arguments{
\item{l_geom}{geometry passed from inside prt_shortpath()}

\item{start_pt}{start point}

\item{end_pt}{end point}
}
\value{
linestring
}
\description{
This function extends the calculated shortest path to start and end with
the provided \emph{start_pt} and \emph{end_pt} geometries. This ensures that the
final sample of new points along the re-routed path include the portion
of the line from the \emph{start_pt}/\emph{end_pt} to the nearest network node.
}
