% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_orient_line.R
\name{prt_orient_line}
\alias{prt_orient_line}
\title{Correctly orient points from line based on a given nearest start point}
\usage{
prt_orient_line(x, l_pts)
}
\arguments{
\item{x}{start point}

\item{l_pts}{points from line}
}
\value{
l_pts
}
\description{
Correctly orient points from line based on a given nearest start point
}
