% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_reroute.R
\name{prt_reroute}
\alias{prt_reroute}
\title{Re-route track points around barrier feature}
\usage{
prt_reroute(trkpts, barrier, vis_graph, blend = TRUE)
}
\arguments{
\item{trkpts}{Simple Feature points ('sf', 'sfc_POINT'/'sfc_MULTIPOINT') that represent
track points. Order is accepted as is and the bounding box of trkpts should be within
the bounding box of the barrier polygon.}

\item{barrier}{Simple Feature polygon ('sf', 'sfc_POLYGON'/'sfc_MULTIPOLYGON')
representing the barrier feature. Should be the same barrier as supplied to the
\code{prt_visgraph()} function.}

\item{vis_graph}{sfnetwork from prt_visgraph()}

\item{blend}{boolean whether to blend start/end points into network}
}
\value{
a two-column tibble with column \emph{fid} representing the row index in trkpts
to be replaced by the new geometry in \emph{geometry} column. If trkpts and barrier do not
spatially intersect and empty tibble is returned.
}
\description{
This is a convenience wrapper, and the suggested function, for re-routing a \emph{trkpts}
series of ordered POINT features around a \emph{barrier} polygon via \emph{vis_graph} built with
the \code{prt_visgraph()} function. The output can be used as a starting point for a custom
process to replace the original geometry. Or, provide the output tibble directly to
\code{prt_update_points()} along with \emph{trkpts} for simply updating in place.
}
\details{
The \code{blend = TRUE} argument will blend all of the \emph{start_pt} / \emph{end_pt}
geometries in \emph{segs_tbl} into the \emph{vis_graph} network via the
\code{sfnetworks::st_network_blend()} function. This process creates new nodes
within \emph{vis_graph} that are positioned at the perpendicular intersection
between each point and the nearest edge. With \code{blend = FALSE} the nearest
existing node is used. In highly complex coastlines, the use of \code{blend = FALSE}
could result in re-routed paths that still intersect land or do not
accurately represent the intended result. For less complex situations and
when computational speed is important, \code{blend = FALSE} may be appropriate.
}
