% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_segment_routes.R
\name{prt_segment_routes}
\alias{prt_segment_routes}
\title{Re-route segments that cross barrier}
\usage{
prt_segment_routes(segs_tbl, sln)
}
\arguments{
\item{segs_tbl}{tbl from get_barrier_segments}

\item{sln}{SpatialLinesNetwork}
}
\value{
data frame
}
\description{
Re-route segments that cross barrier
}
