% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_trim.R
\name{prt_trim}
\alias{prt_trim}
\title{Trim tracks to start and end outside barrier}
\usage{
prt_trim(trkpts, barrier)
}
\arguments{
\item{trkpts}{Simple Feature points ('sf', 'sfc_POINT'/'sfc_MULTIPOINT') that represent
track points. Order is accepted as is and the bounding box of trkpts should be within
the bounding box of the barrier polygon.}

\item{barrier}{Simple Feature polygon ('sf', 'sfc_POLYGON'/'sfc_MULTIPOLYGON')
representing the barrier feature. Should be the same barrier as supplied to the
\code{prt_visgraph()} function.}
}
\description{
Trim tracks to start and end outside barrier
}
