% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_update_points.R
\name{prt_update_points}
\alias{prt_update_points}
\title{Update track points with fixed geometry}
\usage{
prt_update_points(track_pts, segs_tbl)
}
\arguments{
\item{track_pts}{point geometry}

\item{segs_tbl}{segments table}
}
\value{
point geometry
}
\description{
Update track points with fixed geometry
}
