% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_predicates.R
\name{spatial_predicates}
\alias{spatial_predicates}
\alias{not_crosses}
\alias{not_within}
\alias{not_intersects}
\title{Spatial predicates}
\usage{
not_crosses(x, y)

not_within(x, y)

not_intersects(x, y)
}
\arguments{
\item{x, y}{simple features.}
}
\description{
These are custom spatial predicate functions that are negated versions
of the spatial predicates \code{st_within()}, \code{st-crosses()}, and \code{st_intersects}
}
