% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_graph.R
\docType{data}
\name{vis_graph}
\alias{vis_graph}
\title{visual graph for Alaska coast region}
\format{
sfNetwork
}
\usage{
vis_graph
}
\description{
A visual graph built with \code{prt_visgraph()} and the Alaska coastal region used
by the harbor seal example. Included to save time during vignette build.
}
\keyword{datasets}
