% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_barrier_segments.R
\name{get_barrier_segments}
\alias{get_barrier_segments}
\title{Identify Track Points That Intersect With A Barrier}
\usage{
get_barrier_segments(trkpts, barrier)
}
\arguments{
\item{trkpts}{simple feature points that represent track}

\item{barrier}{simple feature polygon for barrier}
}
\value{
data.frame of on-barrier points by segment
}
\description{
Identify Track Points That Intersect With A Barrier
}
