% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_extend_line.R
\name{prt_extend_line}
\alias{prt_extend_line}
\title{Extend a linestring to include given start and end points}
\usage{
prt_extend_line(l, start_pt, end_pt)
}
\arguments{
\item{l}{linestring}

\item{start_pt}{start point}

\item{end_pt}{end point}
}
\value{
linestring
}
\description{
Extend a linestring to include given start and end points
}
