% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_extend_path.R
\name{prt_extend_path}
\alias{prt_extend_path}
\title{Extend a path to include given start and end points}
\usage{
prt_extend_path(l_geom, start_pt, end_pt)
}
\arguments{
\item{start_pt}{start point}

\item{end_pt}{end point}

\item{p}{sfnetwork representing a single path}
}
\value{
linestring
}
\description{
Extend a path to include given start and end points
}
