% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_nearestnode.R
\name{prt_nearestnode}
\alias{prt_nearestnode}
\title{Find the nearest node for start and end points in segs_tbl}
\usage{
prt_nearestnode(segs_tbl, vis_graph)
}
\arguments{
\item{segs_tbl}{output from \code{get_barrier_segments()}}

\item{vis_graph}{SpatialLinesNetwork output from \code{prt_visgraph()}}
}
\value{
data frame with updated columns for nearest start and end nodes
}
\description{
Find the nearest node for start and end points in segs_tbl
}
