% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_reroute.R
\name{prt_reroute}
\alias{prt_reroute}
\title{Re-route track points around barrier feature}
\usage{
prt_reroute(trkpts, barrier, vis_graph)
}
\arguments{
\item{trkpts}{Simple Feature points ('sf', 'sfc_POINT'/'sfc_MULTIPOINT') that represent
track points. Order is accepted as is and the bounding box of trkpts should be within
the bounding box of the barrier polygon.}

\item{barrier}{Simple Feature polygon ('sf', 'sfc_POLYGON'/'sfc_MULTIPOLYGON')
representing the barrier feature. Should be the same barrier as supplied to the
\code{prt_visgraph()} function.}

\item{vis_graph}{SpatialLinesNetwork from prt_visgraph()}
}
\value{
a two-column tibble with column \emph{fid} representing the row index in trkpts
to be replaced by the new geometry in \emph{geometry} column. If trkpts and barrier do not
spatially intersect and empty tibble is returned.
}
\description{
This is a convenience wrapper, and the suggested function, for re-routing a \emph{trkpts}
series of ordered POINT features around a \emph{barrier} polygon via \emph{vis_graph} built with
the \code{prt_visgraph()} function. The output can be used as a starting point for a custom
process to replace the original geometry. Or, provide the output tibble directly to
\code{prt_update_points()} along with \emph{trkpts} for simply updating in place.
}
