% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_shortpath.R
\name{prt_shortpath}
\alias{prt_shortpath}
\title{Calculate the shortest path through a visibility network between two points}
\usage{
prt_shortpath(segs_tbl, vis_graph)
}
\arguments{
\item{segs_tbl}{tbl from get_barrier_segments()}

\item{vis_graph}{SpatialLinesNetwork from prt_visgraph()}
}
\value{
data frame seg_tbl with added geometry column for shortest path LINESTRING
}
\description{
Calculate the shortest path through a visibility network between two points
}
