% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_shortpath.R
\name{prt_shortpath}
\alias{prt_shortpath}
\title{Calculate the shortest path through a visibility network between two points}
\usage{
prt_shortpath(segs_tbl, vis_graph, blend = TRUE)
}
\arguments{
\item{segs_tbl}{tbl from \code{get_barrier_segments()}}

\item{vis_graph}{sfnetwork from prt_visgraph()}

\item{blend}{boolean whether to blend start/end points into network}
}
\value{
segs_tbl data frame with added geometry column for shortest path LINESTRING that
connects the \emph{start_pt} and \emph{end_pt} coordinates
}
\description{
Calculate the shortest path through a visibility network between two points
}
