% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_update_points.R
\name{prt_update_points}
\alias{prt_update_points}
\title{Update track points with fixed geometry}
\usage{
prt_update_points(rrt_pts, trkpts)
}
\arguments{
\item{rrt_pts}{output from \code{prt_reroute()} or tibble with \emph{rrt_idx} and \emph{geometry} columns}

\item{trkpts}{original trkpts Simple Features Collection}
}
\value{
trkpts with updated geometry
}
\description{
Original geometry is updated in place and (currently) no record of those points
that were updated is provided.
}
