% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_update_points.R
\name{prt_update_points}
\alias{prt_update_points}
\title{Update track points with fixed geometry}
\usage{
prt_update_points(track_pts, segs_tbl)
}
\arguments{
\item{track_pts}{original trkpts Simple Features Collection}

\item{segs_tbl}{output from \code{prt_shortpath()}}
}
\value{
trackpts with updated geometry
}
\description{
Original geometry is updated in place and (currently) no record of those points
that were updated is provided.
}
