% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prt_visgraph.R
\name{prt_visgraph}
\alias{prt_visgraph}
\title{Create a visibility graph}
\usage{
prt_visgraph(
  barrier,
  buffer = 0,
  centroids = FALSE,
  centroid_limit = 1e+07,
  aug_points = NULL
)
}
\arguments{
\item{barrier}{simple feature 'POLYGON' or 'MULTIPOLYGON' that can be cast into 'POLYGON'}

\item{buffer}{integer specifying buffer distance for barrier}

\item{centroids}{logical whether to include centroids in the mesh}

\item{centroid_limit}{integer minimum size (m^2) for adding centroid to triangles}

\item{aug_points}{simple feature 'POINT' or 'MULTIPOINT'; Not implemented yet}
}
\value{
SpatialLinesNetwork
}
\description{
Create a visibility graph
}
